Updated GroupCard
